fix(build): TASK-P0-1 Linux 平台编译修复（gcc 13 兼容性）
Ubuntu 24.04 + gcc 13.3.0 + cmake 3.28.3 验证通过：
- 编译零错误零警告
- 单元测试 336 cases / 1142 assertions 全部通过

修复内容：

1. C++ 默认参数解析（gcc 13 严格性）
   - react_loop.h: 委托构造替代 Config = {} 默认参数
   - context_compressor.h: 同上，添加无参构造重载
   - terminal.h: = {} 改为 = TerminalConfig{}

2. designated initializer 缺字段（-Wmissing-field-initializers）
   - result.h: ToolResult 工厂方法补全 .data/.is_error/.text
   - inclaude/types.h: CommandResult 工厂方法补全字段
   - parser.h: ParsedInput 返回值补全 5 字段
   - processor.h: ProcessResult 返回值补全 4 字段
   - chat_session.cpp: BackendStatusEvent 补全 .error
   - test_config_manager.cpp: ConfigMeta 补全 .is_required/.change_callback

3. Linux 平台缺失头文件
   - encoding.cpp: Linux 分支补 <iconv.h> + <cerrno>
   - markdown_renderer.cpp: 补 <cstring>（std::strlen）

4. 死代码与未使用参数
   - file_edit_tool.cpp: 删除未使用的 read_file_raw
   - markdown_renderer.cpp: max_width 加 (void) 标记
   - test_thread_pool.cpp: int vs size_t 符号比较，加强制转换
   - test_event_bus.cpp: t 加 [[maybe_unused]]

5. .gitignore 补充 build-wsl/

+++ tests/unit/core/task/test_thread_pool.cpp
@@ -140,7 +140,7 @@ TEST_CASE("ThreadPool limits concurrency to worker count", "[thread_pool][concur
     }
 
     // 并发峰值不应超过 worker 数（允许瞬时统计偏差，加 1 容差）
-    REQUIRE(max_concurrent.load() <= pool.worker_count() + 1);
+    REQUIRE(static_cast<size_t>(max_concurrent.load()) <= pool.worker_count() + 1);
     REQUIRE(max_concurrent.load() >= 2);  // 至少发生过并发
 }
